+++ forui/lib/src/widgets/tooltip.dart
@@ -1,3 +1,4 @@
+import 'dart:collection';
 import 'dart:ui';
 
 import 'package:flutter/foundation.dart';
@@ -249,6 +250,7 @@ class FTooltip extends StatefulWidget {
 }
 
 class _FTooltipState extends State<FTooltip> with SingleTickerProviderStateMixin {
+  final FocusNode _focus = FocusNode(debugLabel: 'FTooltip', canRequestFocus: false, skipTraversal: true);
   late FTooltipController _controller = widget.controller ?? FTooltipController(vsync: this);
   int _monotonic = 0;
 
@@ -269,6 +271,7 @@ class _FTooltipState extends State<FTooltip> with SingleTickerProviderStateMixin
     if (widget.controller == null) {
       _controller.dispose();
     }
+    _focus.dispose();
     super.dispose();
   }
 
@@ -281,7 +284,16 @@ class _FTooltipState extends State<FTooltip> with SingleTickerProviderStateMixin
     if (widget.hover || widget.longPress) {
       child = CallbackShortcuts(
         bindings: {const SingleActivator(LogicalKeyboardKey.escape): _exit},
-        child: Focus(onFocusChange: (focused) async => focused ? _enter() : _exit(), child: child),
+        child: Focus(
+          // This is required as onFocusChange is not called when focus is shifted from a child to a nested child.
+          onKeyEvent: (_, event) {
+            _toggle(_focus.hasFocus);
+            return KeyEventResult.ignored;
+          },
+          focusNode: _focus,
+          onFocusChange: _toggle,
+          child: child,
+        ),
       );
     }
 
@@ -359,6 +371,26 @@ class _FTooltipState extends State<FTooltip> with SingleTickerProviderStateMixin
         child: child,
       ),
     );
+  }
+
+  Future<void> _toggle(bool focused) {
+    if (!focused) {
+      return _exit();
+    }
+
+    final descendants = Queue.of(_focus.children);
+    while (descendants.isNotEmpty) {
+      final current = descendants.removeFirst();
+      if (current.hasPrimaryFocus) {
+        return _enter();
+      }
+
+      if (!current.canRequestFocus) {
+        descendants.addAll(current.children);
+      }
+    }
+
+    return _exit();
   }
 
   Future<void> _enter() async {